Name things better #80
Our X2Connector is doing import work too so we should call it X2Importer
We're going to have multiple SFTP sites going so let's prefix the local env vars for each one
