+++ app/importers/attendance_importer.rb
@@ -1,5 +1,5 @@
 class AttendanceImporter
-	include X2Connector
+	include X2Importer
 
 	def export_file_name
 		'attendance_export.txt'
